requirements.txt
fastapi[standard]
uvicorn[standard]
pyspice
python-multipart
pydantic
numpy
scipy
matplotlib
pandas

# Database dependencies (MongoDB)
motor
pymongo
dnspython

# Authentication dependencies  
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv

# Additional utilities
email-validator

# Additional utilities
python-dotenv

